+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [0.13.0](https://github.com/flaport/sax/compare/0.12.2...0.13.0)
 
 - Deprecate `sax.nn`.
-- Remove support for pydantic v1.
+- Deprecate support for pydantic v1.
 - Add support for gdsfactory 8 netlists (i.e. `nets` in stead of `connections`)
 
 ## [0.10.2](https://github.com/flaport/sax/compare/0.10.1...0.10.2)